+++ Gemfile
@@ -1,0 +1,6 @@
+source 'https://rubygems.org'
+
+gem 'libxml-ruby'
+
+# Specify your gem's dependencies in wiki_bacon.gemspec
+gemspec
