+++ lib/wiki_bacon/version.rb
@@ -1,0 +1,3 @@
+module WikiBacon
+  VERSION = "0.0.1"
+end
